requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
PyYAML==6.0.1
python-dateutil==2.8.2
urllib3==2.0.7
Flask==2.3.3
Werkzeug==2.3.7
stripe>=8.0.0
rq>=1.15.0
redis>=4.5.0 